be/service/requirements.txt
aniso8601==9.0.1
attrs==21.4.0
bandit==1.7.4
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
docx==0.2.4
docx2pdf==0.1.8
docxcompose==1.3.4
Flask==2.1.2
Flask-Cors==3.0.10
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.8.1
google-api-python-client==2.50.0
google-auth==2.7.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.2
greenlet==1.1.2
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.6.0
lxml==4.9.0
MarkupSafe==2.1.1
naver-config==0.0.2
naver-core==0.0.3
naver-db==0.0.2
naver-net==0.0.2
numpy==1.22.4
oauth2client==4.1.3
oauthlib==3.2.0
pandas==1.4.2
pathlib==1.0.1
pbr==5.9.0
Pillow==9.1.1
protobuf==3.20.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pyrsistent==0.18.1
python-consul==1.1.0
python-dateutil==2.8.2
python-decouple==3.6
python-docx==0.8.11
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.37
stevedore==3.5.0
tqdm==4.64.0
uritemplate==4.1.1
urllib3==1.26.9
waitress==2.1.2
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
